Add member data access to window classes

+++ Nautilus/core/Types.h
@@ -21,8 +21,8 @@ namespace nt
     using uint8   = unsigned char;
     using int16   = short;
     using uint16  = unsigned short;
-    using int32   = long;
-    using uint32  = unsigned long;
+    using int32   = int;
+    using uint32  = unsigned int;
     using int64   = long long;
     using uint64  = unsigned long long;
 #if defined(__int128)
